Allow for serial device name to be passed on init.
Thanks goes to kalhamaar for suggesting this change.  The default serial device is still '/dev/ttyACM0', but now you can provide the device name, rather than a port number, which provides better flexibility to work with other operating systems or system configurations.  For Windows a OS, for instance, you would use 'COM3', or whatever is the assigned port, when instancing a Controller object.

Also kalhamaar noted a bug in the getMax function, which has also been corrected.